Move SVG utility to it's own class
We had 3 places reading/converting SVGs, with multiple copies of the image
transcoder and parsing w/h info. This moves it all to a shared SVGUtil class
with only 2 public functions: loading SVGs (2 versions) and converting to
BufferedImages.

Since all the work is now done in the model jar the batik jar and its
dependencies no longer need to be exported.

Some minor formatting and variable overriding cleanup was automatically
applied to the image generator.
