+++ ContestModel/.classpath
@@ -16,9 +16,9 @@
 			<attribute name="org.eclipse.jst.component.dependency" value="../"/>
 		</attributes>
 	</classpathentry>
-	<classpathentry exported="true" kind="lib" path="lib/xml-apis-ext-1.3.04.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/xmlgraphics-commons-2.11.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/batik-all-1.19.jar"/>
+	<classpathentry kind="lib" path="lib/xml-apis-ext-1.3.04.jar"/>
+	<classpathentry kind="lib" path="lib/xmlgraphics-commons-2.11.jar"/>
+	<classpathentry kind="lib" path="lib/batik-all-1.19.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/sentry-opentelemetry-agent-8.12.0.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
